+++ CHANGELOG.md
@@ -36,6 +36,8 @@ be enable with an 2nd Optional Boolean Argument as `true`
 - added Tests for join paths and other paths function
 - added a Makefile for the build so i dont have to remember commands
 - added `dir` tests
+- added function to convert compressed json in pretty json and the other way around
+- added tests for JSON
 
 Read the DOCS for Version 0.2.0 for more Infos